Created HttpRequestUtils Class to Take Over Http Request
Such as the title.
